+++ ui/components/checkbox/index.stories.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import * as CheckboxExamples from './base/example';
+import * as CheckboxFormElementExamples from './form-element/example';
+import { getAllDisplayCollectionsByType } from '../../shared/helpers';
+import StoryFrame from '../../../shared/components/StoryFrame';
+
+import '../../index.scss';
+
+const examples = [CheckboxExamples, CheckboxFormElementExamples];
+
+const kitchenSink = getAllDisplayCollectionsByType(examples, [
+  'default',
+  'examples',
+  'states'
+]);
+
+storiesOf('Components/Checkbox', module).add('Kitchen Sink', () =>
+  kitchenSink.map((element, idx) =>
+    element.map(({ label, component }) => (
+      <StoryFrame
+        component={component}
+        label={label}
+        key={`kitchen-sink-${label}-${idx}`}
+      />
+    ))
+  )
+);