+++ ui/components/menus/submenu/index.stories.js
@@ -1,0 +1,9 @@
+import { storiesOf } from '@storybook/react';
+import * as SubmenuExamples from './example';
+
+import '../../../index.scss';
+
+storiesOf('Components/Menus/Submenu', module).add(
+  'Base',
+  () => SubmenuExamples.default
+);