+++ ui/components/page-headers/base/index.stories.js
@@ -1,0 +1,9 @@
+import { storiesOf } from '@storybook/react';
+import * as BaseExamples from './example';
+
+import '../../../index.scss';
+
+storiesOf('Components/Page Headers/Base', module).add(
+  'Base',
+  () => BaseExamples.default
+);